Move functions to separate file, update functions
